QP estimator is kind of working
We replaced the OSD equality constraint with the inverse OSD constraint,
where the inverse of the Lambda matrix is used to avoid potential
numerical issues induced by matrix inversion.

The end-effector velocity jump is also predicted. They look OK.

The joint velocity jump is on the same level as the analytical approach.

The result is improved. However we can not see clearly if it performs
better than the analytical solution.

Maybe it worth tying the Laranger multiplier method.